refactor(notifications): move show_and_auto_close out of app.rs
