+++ cosmic-ext-connected/src/notifications.rs
@@ -118,6 +118,37 @@ fn should_show_notification(dedup_file: &str, key: &str) -> bool {
     should_notify
 }
 
+/// Show a notification and auto-close it after the configured timeout.
+///
+/// COSMIC's notification daemon does not respect the `expire_timeout` hint from
+/// the freedesktop notification spec, so all notifications display for the daemon's
+/// fixed duration regardless of the value passed. To work around this, notifications
+/// are created with `Timeout::Never` (expire_timeout=0) to prevent the daemon from
+/// auto-closing them, then explicitly closed after the user's configured timeout.
+pub async fn show_and_auto_close(
+    notification: notify_rust::Notification,
+    timeout_ms: u32,
+    label: &str,
+) {
+    let label = label.to_string();
+    let result = tokio::task::spawn_blocking(move || notification.show()).await;
+    match result {
+        Ok(Ok(handle)) => {
+            tokio::spawn(async move {
+                tokio::time::sleep(std::time::Duration::from_millis(timeout_ms as u64)).await;
+                // close() uses zbus::block_on internally, so run in blocking context
+                let _ = tokio::task::spawn_blocking(move || handle.close()).await;
+            });
+        }
+        Ok(Err(e)) => {
+            tracing::warn!("Failed to show {} notification: {}", label, e);
+        }
+        Err(e) => {
+            tracing::warn!("Notification task panicked: {}", e);
+        }
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
